# docs/api-documentations/openapi/asset.openapi.json
{
  "components": {
    "schemas": {
...
      "BatchCorrectionDto": {
        "properties": {
          "brandId": {
            "description": "品牌（字典项 id）",
            "minimum": 1,
            "type": "number"
          },
          "reason": {
            "description": "纠正原因（必填）",
            "maxLength": 500,
            "type": "string"
          },
          "remark": {
            "description": "批次备注",
            "maxLength": 500,
            "type": "string"
          },
          "spec": {
            "description": "规格（可空：不纠正规格；非空且无后续流水时可纠正）",
            "maxLength": 100,
            "type": "string"
          },
          "supplierId": {
            "description": "供应商（字典项 id）",
            "minimum": 1,
            "type": "number"
          },
          "unitPrice": {
            "description": "单价（元，最多两位小数）",
            "example": "10.00",
            "type": "string"
          },
          "warehouseId": {
            "description": "目标库位 id（可空：不移动库位；非空且无后续流水时可纠正）",
            "minimum": 1,
            "type": "number"
          }
        },
        "required": [
          "reason"
        ],
        "type": "object"
      },
...
      "QrActionDto": {
        "properties": {
          "action": {
            "description": "动作",
            "enum": [
              "DISABLE",
              "ENABLE",
              "REGENERATE"
            ],
            "type": "string"
          }
        },
        "required": [
          "action"
        ],
        "type": "object"
      },
      "QrCodeCreateDto": {
        "properties": {
          "targetId": {
            "description": "目标标识（SCAN_CATALOG 为 null）",
            "minimum": 1,
            "type": "number"
          },
          "targetType": {
            "description": "目标类型",
            "enum": [
              "ASSET",
              "INVENTORY_ITEM",
              "SCAN_CATALOG"
            ],
            "type": "string"
          }
        },
        "required": [
          "targetType"
        ],
        "type": "object"
      },
      "QrParseDto": {
        "properties": {
          "publicId": {
            "description": "二维码公开标识（来自 /scan#<publicId> fragment；解析后立即从地址栏移除）",
            "maxLength": 200,
            "type": "string"
          }
        },
        "required": [
          "publicId"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "csrfHeader": {
        "description": "CSRF 双提交头（状态变更请求必须与 wbme_csrf Cookie 一致）",
        "in": "header",
        "name": "X-WBME-CSRF-Token",
        "type": "apiKey"
      },
      "sessionCookie": {
        "description": "服务端会话 Cookie（HttpOnly；登录/激活成功后下发）",
        "in": "cookie",
        "name": "wbme_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "asset 部署单元（资产系统）OpenAPI 契约。构建期生成（主 PRD §9.5）：统一错误结构见 components.schemas.ErrorResponse；默认会话 Cookie + CSRF 头认证，公开路由无 security。",
    "title": "WBME asset API",
    "version": "0.1.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/inventory/batches": {
      "get": {
        "operationId": "InventoryController_listBatches",
        "parameters": [
          {
            "description": "页码（从 1 开始）",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "description": "每页条数（服务端接受 10/20/50/100）",
            "in": "query",
            "name": "pageSize",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 10,
              "type": "number"
            }
          },
          {
            "description": "库存条目 id",
            "in": "query",
            "name": "inventoryItemId",
            "required": false,
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "description": "品种 id",
            "in": "query",
            "name": "consumableId",
            "required": false,
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "description": "库位 id",
            "in": "query",
            "name": "warehouseId",
            "required": false,
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Inventory"
        ]
      }
    },
    "/api/v1/inventory/batches/{id}/corrections": {
      "post": {
        "operationId": "InventoryController_correctBatch",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchCorrectionDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Inventory"
        ]
      }
    },
    "/api/v1/inventory/items": {
      "get": {
        "operationId": "InventoryController_listItems",
        "parameters": [
          {
            "description": "页码（从 1 开始）",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "description": "每页条数（服务端接受 10/20/50/100）",
            "in": "query",
            "name": "pageSize",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 10,
              "type": "number"
            }
          },
          {
            "description": "品种 id",
            "in": "query",
            "name": "consumableId",
            "required": false,
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "description": "库位 id",
            "in": "query",
            "name": "warehouseId",
            "required": false,
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "description": "规格（精确）",
            "in": "query",
            "name": "spec",
            "required": false,
            "schema": {
              "maxLength": 100,
              "type": "string"
            }
          },
          {
            "description": "仅显示低库存品种（可用 < 安全库存）",
            "in": "query",
            "name": "lowStockOnly",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Inventory"
        ]
      }
    },
    "/api/v1/inventory/stock-flows": {
      "get": {
        "operationId": "InventoryController_listStockFlows",
        "parameters": [
          {
            "description": "页码（从 1 开始）",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "description": "每页条数（服务端接受 10/20/50/100）",
            "in": "query",
            "name": "pageSize",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 10,
              "type": "number"
            }
          },
          {
            "description": "品种 id",
            "in": "query",
            "name": "consumableId",
            "required": false,
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "description": "库存条目 id",
            "in": "query",
            "name": "inventoryItemId",
            "required": false,
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "description": "库位 id",
            "in": "query",
            "name": "warehouseId",
            "required": false,
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "description": "流水类型",
            "in": "query",
            "name": "flowType",
            "required": false,
            "schema": {
              "enum": [
                "STOCK_IN",
                "ISSUE",
                "DEDUCTION",
                "RETURN",
                "TRANSFER_OUT",
                "TRANSFER_IN",
                "CORRECTION"
              ],
              "type": "string"
            }
          },
          {
            "description": "业务来源类型（申请/调拨/处置等）",
            "in": "query",
            "name": "refType",
            "required": false,
            "schema": {
              "maxLength": 40,
              "type": "string"
            }
          },
          {
            "description": "业务来源标识",
            "in": "query",
            "name": "refId",
            "required": false,
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Inventory"
        ]
      }
    },
    "/api/v1/inventory/stock-flows/export": {
      "get": {
        "operationId": "InventoryController_exportStockFlows",
        "parameters": [
          {
            "description": "页码（从 1 开始）",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "description": "每页条数（服务端接受 10/20/50/100）",
            "in": "query",
            "name": "pageSize",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 10,
              "type": "number"
            }
          },
          {
            "description": "品种 id",
            "in": "query",
            "name": "consumableId",
            "required": false,
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "description": "库存条目 id",
            "in": "query",
            "name": "inventoryItemId",
            "required": false,
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "description": "库位 id",
            "in": "query",
            "name": "warehouseId",
            "required": false,
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "description": "流水类型",
            "in": "query",
            "name": "flowType",
            "required": false,
            "schema": {
              "enum": [
                "STOCK_IN",
                "ISSUE",
                "DEDUCTION",
                "RETURN",
                "TRANSFER_OUT",
                "TRANSFER_IN",
                "CORRECTION"
              ],
              "type": "string"
            }
          },
          {
            "description": "业务来源类型（申请/调拨/处置等）",
            "in": "query",
            "name": "refType",
            "required": false,
            "schema": {
              "maxLength": 40,
              "type": "string"
            }
          },
          {
            "description": "业务来源标识",
            "in": "query",
            "name": "refId",
            "required": false,
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Inventory"
        ]
      }
    },
...
    "/api/v1/qr-codes": {
      "get": {
        "operationId": "QrController_list",
        "parameters": [
          {
            "description": "页码（从 1 开始）",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "description": "每页条数（服务端接受 10/20/50/100）",
            "in": "query",
            "name": "pageSize",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 10,
              "type": "number"
            }
          },
          {
            "description": "目标类型",
            "in": "query",
            "name": "targetType",
            "required": false,
            "schema": {
              "enum": [
                "ASSET",
                "INVENTORY_ITEM",
                "SCAN_CATALOG"
              ],
              "type": "string"
            }
          },
          {
            "description": "状态",
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "enum": [
                "ACTIVE",
                "DISABLED",
                "REVOKED"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Qr"
        ]
      },
      "post": {
        "operationId": "QrController_create",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QrCodeCreateDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Qr"
        ]
      }
    },
    "/api/v1/qr-codes/{id}/action": {
      "post": {
        "operationId": "QrController_action",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QrActionDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Qr"
        ]
      }
    },
    "/api/v1/qr-codes/parse": {
      "post": {
        "operationId": "QrController_parse",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QrParseDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Qr"
        ]
      }
    },
...
  },
  "security": [
    {
      "csrfHeader": [],
      "sessionCookie": []
    }
  ],
  "servers": [],
...
}